update sight type

+++ game/floor/floor_actturn.go
@@ -733,11 +733,11 @@ func (f *Floor) sendViewportNoti(
 		aox, aoy = f.terrain.WrapXY(aox, aoy)
 		vpixyolists := vpixyolistcache.GetAtByCache(aox, aoy)
 		sightMat := f.terrain.GetViewportCache().GetByCache(aox, aoy)
-		sight := ao.GetTurnData().Sight
-		aOs := f.makeViewportActiveObjs2(vpixyolists[0], sightMat, float32(sight))
-		pOs := f.makeViewportCarryObjs2(vpixyolists[1], sightMat, float32(sight))
-		fOs := f.makeViewportFieldObjs2(vpixyolists[2], sightMat, float32(sight))
-		dOs := f.makeViewportDangerObjs2(vpixyolists[3], sightMat, float32(sight))
+		sight := float32(ao.GetTurnData().Sight)
+		aOs := f.makeViewportActiveObjs2(vpixyolists[0], sightMat, sight)
+		pOs := f.makeViewportCarryObjs2(vpixyolists[1], sightMat, sight)
+		fOs := f.makeViewportFieldObjs2(vpixyolists[2], sightMat, sight)
+		dOs := f.makeViewportDangerObjs2(vpixyolists[3], sightMat, sight)
 		// update ai floor4client info
 		f4c := ao.GetFloor4Client(f.GetName())
 		for _, fo := range fOs {
@@ -746,10 +746,10 @@ func (f *Floor) sendViewportNoti(
 
 		if ao.GetAndClearNeedTANoti() {
 			ao.UpdateVisitAreaBySightMat2(f, aox, aoy, sightMat,
-				float32(sight))
+				sight)
 			if aoconn := ao.GetClientConn(); aoconn != nil {
 				// make and send NotiTileArea
-				cstiles := f.makeViewportTiles2(aox, aoy, sightMat, float32(sight))
+				cstiles := f.makeViewportTiles2(aox, aoy, sightMat, sight)
 				notiTA := &c2t_obj.NotiVPTiles_data{
 					FloorName: f.GetName(),
 					VPX:       aox,
